3 tests migrated
