+++ crates/burn-jit/src/fusion/elemwise/optimization.rs
@@ -193,181 +193,3 @@ impl<R: Runtime> ElementWise<R, ExecutionPhase<R>> {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::tests::TestRuntime;
-    use crate::JitBackend;
-
-    use burn_fusion::stream::Operation;
-    use burn_fusion::{Fusion, FusionBackend};
-    use burn_tensor::Int;
-    use burn_tensor::{backend::Backend, Data, Tensor};
-
-    #[test]
-    fn test_fusion_same_behavior() {
-        type Backend = JitBackend<TestRuntime>;
-        type FusedBackend = Fusion<Backend>;
-
-        let data_1 = Tensor::<FusedBackend, 2>::random(
-            [1, 32],
-            burn_tensor::Distribution::Default,
-            &Default::default(),
-        )
-        .into_data();
-        let data_2 = Tensor::<Backend, 2>::random(
-            [32, 32],
-            burn_tensor::Distribution::Default,
-            &Default::default(),
-        )
-        .into_data();
-
-        let result_ref = execute::<Backend>(
-            data_1.clone(),
-            data_2.clone(),
-            ImplementationDetails::Variant1,
-        );
-        let result_fused = execute::<FusedBackend>(
-            data_1.clone(),
-            data_2.clone(),
-            ImplementationDetails::Variant1,
-        );
-
-        result_ref.assert_approx_eq(&result_fused, 3);
-    }
-
-    #[test]
-    fn test_fusion_same_behavior_int() {
-        let data_1 = Tensor::<FusedBackend, 2>::random(
-            [32, 32],
-            burn_tensor::Distribution::Default,
-            &Default::default(),
-        )
-        .into_data();
-        let data_2 = Tensor::<Backend, 2>::random(
-            [32, 32],
-            burn_tensor::Distribution::Default,
-            &Default::default(),
-        )
-        .into_data()
-        .convert();
-
-        fn func<B: burn_tensor::backend::Backend>(
-            data1: Data<f32, 2>,
-            data2: Data<i32, 2>,
-        ) -> Data<f32, 2> {
-            let x = Tensor::<B, 2>::from_data(data1.convert(), &Default::default());
-            let y = Tensor::<B, 2, Int>::from_data(data2.convert(), &Default::default());
-
-            let x_1 = x.clone().powf_scalar(2.0);
-            let x_1 = x_1 + x;
-            let y_1 = y * 6;
-            let y_1 = y_1 + 4;
-
-            let z = x_1 * y_1.float();
-
-            z.into_data().convert()
-        }
-
-        type Backend = JitBackend<TestRuntime>;
-        type FusedBackend = Fusion<Backend>;
-
-        let result_fused = func::<FusedBackend>(data_1.clone(), data_2.clone());
-        let result_ref = func::<Backend>(data_1.clone(), data_2.clone());
-
-        result_ref.assert_approx_eq(&result_fused, 3);
-    }
-
-    #[test]
-    fn test_fusion_same_behavior_different_variant() {
-        type Backend = JitBackend<TestRuntime>;
-        type FusedBackend = Fusion<Backend>;
-
-        let data_1 = Tensor::<FusedBackend, 2>::random(
-            [1, 32],
-            burn_tensor::Distribution::Default,
-            &Default::default(),
-        )
-        .into_data();
-        let data_2 = Tensor::<Backend, 2>::random(
-            [32, 32],
-            burn_tensor::Distribution::Default,
-            &Default::default(),
-        )
-        .into_data();
-
-        let result_ref = execute::<Backend>(
-            data_1.clone(),
-            data_2.clone(),
-            ImplementationDetails::Variant2,
-        );
-        let result_fused_variant1 = execute::<FusedBackend>(
-            data_1.clone(),
-            data_2.clone(),
-            ImplementationDetails::Variant1,
-        );
-        let result_fused_variant2 = execute::<FusedBackend>(
-            data_1.clone(),
-            data_2.clone(),
-            ImplementationDetails::Variant2,
-        );
-
-        result_ref.assert_approx_eq(&result_fused_variant1, 3);
-        result_ref.assert_approx_eq(&result_fused_variant2, 3);
-    }
-
-    #[test]
-    fn test_end_condition_scalar_ops() {
-        type Backend = JitBackend<TestRuntime>;
-        let device = Default::default();
-        let tensor1 = Tensor::<Backend, 2>::ones([32, 32], &device);
-        let tensor2 = Tensor::<Backend, 2>::ones([32, 42], &device);
-        let output = tensor1.exp().log();
-
-        // This will add a scalar to the context, even if the actual operation can't be fused with
-        // the preceding ones because of the shape difference.
-        let _ = tensor2 + 2;
-
-        // When we try to execute the operations, the number of bindings can be different if we are
-        // not careful.
-        Backend::sync(&output.device());
-    }
-
-    struct FakeAddOps;
-
-    impl<B: FusionBackend> Operation<B> for FakeAddOps {
-        fn execute(self: Box<Self>, _: &mut burn_fusion::HandleContainer<B>) {
-            panic!("Should always fused during tests.")
-        }
-    }
-
-    enum ImplementationDetails {
-        Variant1,
-        Variant2,
-    }
-
-    fn execute<B: Backend>(
-        data_1: Data<f32, 2>,
-        data_2: Data<f32, 2>,
-        variant: ImplementationDetails,
-    ) -> Data<f32, 2> {
-        let device = B::Device::default();
-        let tensor_1 = Tensor::<B, 2>::from_data(data_1.convert(), &device);
-        let tensor_2 = Tensor::<B, 2>::from_data(data_2.convert(), &device);
-        let tensor_3 = tensor_1.clone() + tensor_2;
-        let tensor_4 = tensor_3.clone() - tensor_1;
-        let mut tensor_5 = tensor_4.clone() + 5.0;
-        match variant {
-            ImplementationDetails::Variant1 => {}
-            ImplementationDetails::Variant2 => {
-                tensor_5 = tensor_5 + 1;
-                tensor_5 = tensor_5 - 1;
-            }
-        }
-        let tensor_6 = burn_tensor::activation::gelu(tensor_5 + tensor_3.clone());
-        let mask = tensor_4.lower_equal(tensor_3);
-        let tmp = tensor_6.mask_fill(mask, 0.3);
-
-        tmp.into_data().convert()
-    }
-}
