+++ crates/burn-jit/src/kernel/reduce/naive_reduce_shader.rs
@@ -171,126 +171,3 @@ pub fn reduce_dim_naive<
 
     output
 }
-
-#[cfg(feature = "export_tests")]
-mod tests {
-    #[burn_tensor_testgen::testgen(reduction)]
-    mod reduction {
-        use super::*;
-        use burn_jit::kernel::reduce::{sum_dim, ReduceStrategy};
-        use burn_tensor::{ops::IntTensorOps, Data, Distribution, Int, Shape, Tensor};
-
-        #[test]
-        fn reduction_sum_dim_should_work_with_multiple_invocations() {
-            let tensor = Tensor::<TestBackend, 2>::random(
-                [6, 1024],
-                Distribution::Default,
-                &Default::default(),
-            );
-            let tensor_ref =
-                Tensor::<ReferenceBackend, 2>::from_data(tensor.to_data(), &Default::default());
-            let reduce_dim = 1;
-
-            let val =
-                Tensor::<TestBackend, 2>::from_primitive(sum_dim::<TestRuntime, f32, f32, 2>(
-                    tensor.into_primitive(),
-                    reduce_dim,
-                    ReduceStrategy::Naive,
-                ));
-            let val_ref = tensor_ref.sum_dim(1);
-
-            val_ref.into_data().assert_approx_eq(&val.into_data(), 3);
-        }
-
-        // #[test]
-        // fn reduction_argmin_dim_should_work_with_multiple_invocations() {
-        //     let tensor =
-        //         Tensor::<TestBackend, 2>::random([6, 1024], Distribution::Default, &Default::default());
-        //     let tensor_ref =
-        //         Tensor::<ReferenceBackend, 2>::from_data(tensor.to_data(), &Default::default());
-        //     let reduce_dim = 1;
-        //     let output = init_reduce_output(&tensor.clone().into_primitive(), reduce_dim);
-
-        //     let val = Tensor::<TestBackend, 2, Int>::from_primitive(reduce_dim_naive::<
-        //         ArgMin,
-        //         TestRuntime,
-        //         f32,
-        //         i32,
-        //         2,
-        //     >(
-        //         tensor.into_primitive(),
-        //         output,
-        //         reduce_dim,
-        //     ));
-        //     let val_ref = tensor_ref.argmin(reduce_dim);
-
-        //     assert_eq!(val_ref.into_data().convert(), val.into_data());
-        // }
-
-        // #[test]
-        // fn reduction_argmax_dim_should_work_with_multiple_invocations() {
-        //     let tensor =
-        //         Tensor::<TestBackend, 2>::random([6, 1024], Distribution::Default, &Default::default());
-        //     let tensor_ref =
-        //         Tensor::<ReferenceBackend, 2>::from_data(tensor.to_data(), &Default::default());
-        //     let reduce_dim = 1;
-        //     let output = init_reduce_output(&tensor.clone().into_primitive(), reduce_dim);
-
-        //     let val = Tensor::<TestBackend, 2, Int>::from_primitive(reduce_dim_naive::<
-        //         ArgMax,
-        //         TestRuntime,
-        //         f32,
-        //         i32,
-        //         2,
-        //     >(
-        //         tensor.into_primitive(),
-        //         output,
-        //         reduce_dim,
-        //     ));
-        //     let val_ref = tensor_ref.argmax(reduce_dim);
-
-        //     assert_eq!(val_ref.into_data().convert(), val.into_data());
-        // }
-
-        // #[test]
-        // fn sum_dim_should_work_with_int() {
-        //     let summed_shape = Shape::new([1]);
-        //     let data = Data::from([1, 2, 3, 4]);
-        //     let tensor = TestBackend::int_from_data(data, &Default::default());
-
-        //     let summed_tensor = TestBackend::int_empty(summed_shape, &Default::default());
-
-        //     let val = Tensor::<TestBackend, 1, Int>::from_primitive(reduce_dim_naive::<
-        //         SumDim,
-        //         TestRuntime,
-        //         i32,
-        //         i32,
-        //         1,
-        //     >(tensor, summed_tensor, 0));
-
-        //     let sum_as_data = Data::from([10]);
-        //     val.into_data().assert_approx_eq(&sum_as_data, 1);
-        // }
-
-        // #[test]
-        // fn mean_dim_should_work_with_int() {
-        //     let mean_shape = Shape::new([1]);
-        //     let data = Data::from([1, 2, 3, 4]);
-        //     let tensor = TestBackend::int_from_data(data, &Default::default());
-
-        //     let mean_tensor = TestBackend::int_empty(mean_shape, &Default::default());
-
-        //     let val = Tensor::<TestBackend, 1, Int>::from_primitive(reduce_dim_naive::<
-        //         MeanDim,
-        //         TestRuntime,
-        //         i32,
-        //         i32,
-        //         1,
-        //     >(tensor, mean_tensor, 0));
-
-        //     // Mean calculation truncates to an integer
-        //     let mean_as_data = Data::from([2]);
-        //     val.into_data().assert_approx_eq(&mean_as_data, 1);
-        // }
-    }
-}
